src/python_cuda/requirements.txt
sentry-sdk>=1.40.0
cupy-cuda12x>=12.0.0
pycuda>=2023.1
numpy>=1.24.0
psutil>=5.9.0